refactor: dimension 정보 받아오는 함수 분리, 함수명 변경
