+++ src/components/create/StickerInput.tsx
@@ -1,5 +1,5 @@
 'use client';
-import { useGetAllStickers } from '@/hooks/queries/useGetStickers';
+import { useGetCategorizedStickers } from '@/hooks/queries/useGetCategorizedStickers';
 import { useState, useEffect } from 'react';
 import StickerCategoryButton from '@/components/create/stickerInput/StickerCategoryButton';
 import StickerSlot from '@/components/create/stickerInput/StickerSlot';
@@ -10,13 +10,13 @@ import { StickerLoading } from '../loading/StickerLoading';
 const StickerInput = () => {
   const [selectedCategory, setSelectedCategory] = useState<string>('classic');
 
-  const { data: stickerByCategory, isLoading, error } = useGetAllStickers();
+  const { data: categorizedStickers, isLoading, error } = useGetCategorizedStickers();
 
   useEffect(() => {
-    if (stickerByCategory && !stickerByCategory[selectedCategory]) {
+    if (categorizedStickers && !categorizedStickers[selectedCategory]) {
       Notify.info('준비중인 서비스입니다.');
     }
-  }, [selectedCategory, stickerByCategory]);
+  }, [selectedCategory, categorizedStickers]);
 
   if (error) {
     console.error(error.message);
@@ -28,11 +28,11 @@ const StickerInput = () => {
   };
 
   const renderSticker = () => {
-    if (!stickerByCategory) return null;
+    if (!categorizedStickers) return null;
 
     if (selectedCategory === null) {
-      return Object.keys(stickerByCategory).map((category) =>
-        stickerByCategory[category].map((sticker) => (
+      return Object.keys(categorizedStickers).map((category) =>
+        categorizedStickers[category].map((sticker) => (
           <StickerSlot
             key={sticker.id}
             stickerImage={sticker}
@@ -41,7 +41,7 @@ const StickerInput = () => {
       );
     }
 
-    return stickerByCategory[selectedCategory]?.map((sticker) => (
+    return categorizedStickers[selectedCategory]?.map((sticker) => (
       <StickerSlot
         key={sticker.id}
         stickerImage={sticker}
@@ -65,7 +65,7 @@ const StickerInput = () => {
       </div>
       {isLoading ? (
         <StickerLoading />
-      ) : stickerByCategory && stickerByCategory[selectedCategory] ? (
+      ) : categorizedStickers && categorizedStickers[selectedCategory] ? (
         <div className='grid grid-cols-4 gap-[9px] h-[150px] overflow-auto'>{renderSticker()}</div>
       ) : null}
     </div>
