+++ src/utils/image/getImageDimensions.ts
@@ -1,0 +1,10 @@
+const getImageDimensions = (url: string): Promise<{ width: number; height: number }> => {
+  return new Promise((resolve, reject) => {
+    const img = new Image();
+    img.src = url;
+    img.onload = () => resolve({ width: img.width, height: img.height });
+    img.onerror = reject;
+  });
+};
+
+export default getImageDimensions;
